change in form request and migration to table 58

+++ database/migrations/2022_06_23_062759_create_gostaresh_revenue_status_analyses_table.php
@@ -24,14 +24,14 @@ return new class extends Migration
             $table->foreignId('rural_district_id')->nullable();
 
             $table->string('unit')->nullable();
-            $table->unsignedInteger('total_revenue')->comment('کل درآمد ها')->nullable();
-            $table->unsignedInteger('income_from_student_tuition')->comment('درآمد حاصل از شهریه دانشجویان')->nullable();
-            $table->unsignedInteger('income_from_commercialized_technologies')->comment('درصد درآمد حاصل از فروش فناوری و طرح های تجاری سازی شده')->nullable();
-            $table->unsignedInteger('income_from_research_activities')->comment('درصد درآمد حاصل از فعالیت های تحقیق و توسعه واحد')->nullable();
-            $table->unsignedInteger('income_from_skills_training')->comment('درآمدهای حاصل از مهارت آموزی، فعالیت های کاربنیان و کارآفرینی واحد')->nullable();
+            $table->float('total_revenue')->comment('کل درآمد ها')->nullable();
+            $table->float('income_from_student_tuition')->comment('درآمد حاصل از شهریه دانشجویان')->nullable();
+            $table->float('income_from_commercialized_technologies')->comment('درصد درآمد حاصل از فروش فناوری و طرح های تجاری سازی شده')->nullable();
+            $table->float('income_from_research_activities')->comment('درصد درآمد حاصل از فعالیت های تحقیق و توسعه واحد')->nullable();
+            $table->float('income_from_skills_training')->comment('درآمدهای حاصل از مهارت آموزی، فعالیت های کاربنیان و کارآفرینی واحد')->nullable();
             $table->float('operating_income_growth_rate')->comment('نرخ رشد درآمدهای عملیاتی واحد')->nullable();
-            $table->unsignedInteger('total_non_tuition_income')->comment('مجموع درآمدهای غیر شهریه ای واحد')->nullable();
-            $table->unsignedInteger('total_international_income')->comment('مجموع درآمد های ناشی از فعالیت های بین المللی')->nullable();
+            $table->float('total_non_tuition_income')->comment('مجموع درآمدهای غیر شهریه ای واحد')->nullable();
+            $table->float('total_international_income')->comment('مجموع درآمد های ناشی از فعالیت های بین المللی')->nullable();
             $table->float('shareholder_income')->comment('درآمد ناشی از سهامداری')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
